Reduce duplication between McLaren and StoryStream web services tests
