+++ app/src/test/java/com/github/fo2rist/mclaren/web/StoryStreamWebServiceImplTest.kt
@@ -1,50 +1,18 @@
 package com.github.fo2rist.mclaren.web
 
-import com.github.fo2rist.mclaren.testutilities.overrideAnswersToSuccess
-import com.nhaarman.mockitokotlin2.any
-import com.nhaarman.mockitokotlin2.mock
-import com.nhaarman.mockitokotlin2.spy
-import com.nhaarman.mockitokotlin2.verify
-import com.nhaarman.mockitokotlin2.whenever
-import kotlinx.coroutines.runBlocking
-import okhttp3.OkHttpClient
 import org.junit.Before
-import org.junit.Test
 import org.junit.runner.RunWith
-import org.mockito.Mock
-import org.mockito.MockitoAnnotations
 import org.robolectric.RobolectricTestRunner
 
 @RunWith(RobolectricTestRunner::class)
-class StoryStreamWebServiceImplTest {
-    @Mock
-    private lateinit var httpClientMock: OkHttpClient
+class StoryStreamWebServiceImplTest : BaseFeedWebServiceTest() {
 
-    private lateinit var webservice: StoryStreamWebServiceImpl
+    override lateinit var webservice: BaseFeedWebService
 
     @Before
-    fun setUp() {
-        MockitoAnnotations.initMocks(this)
-
-        httpClientMock.overrideAnswersToSuccess()
+    override fun setUp() {
+        super.setUp()
 
         webservice = StoryStreamWebServiceImpl(httpClientMock)
-    }
-
-    @Test
-    fun `test requestLatestFeed requests null page`() = runBlocking<Unit> {
-        val webserviceSpy = spy(webservice)
-        whenever(webserviceSpy.createFeedPageRequest()).thenReturn(mock())
-
-        webserviceSpy.requestLatestFeed()
-
-        verify(webserviceSpy).createFeedPageRequest(null)
-    }
-
-    @Test
-    fun `test requestFeedPage calls http client`() = runBlocking<Unit> {
-        webservice.requestFeedPage(123)
-
-        verify(httpClientMock).newCall(any())
     }
 }
